refactor: move Authenticator into service.util package
